src/api/model/requirements.txt
pandas==2.2.0
joblib==1.3.2
requests==2.31.0
python-dotenv==1.0.1
scikit-learn==1.2.2
fastapi==0.109.2
pydantic==2.6.1
uvicorn==0.27.1
tld==0.13
tldextract==5.1.1
urllib3==2.2.0
xgboost==2.0.3